+++ ft_strmap.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/10 11:22:31 by pholster       #+#    #+#                */
-/*   Updated: 2019/01/31 17:54:37 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/01 12:07:07 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_strmap(const char *s, char (*f)(char))
 	i = 0;
 	if (f == NULL || s == NULL)
 		return (NULL);
-	len = ft_strlen(s);
+	len = (int)ft_strlen(s);
 	str = ft_strnew(len);
 	if (str == NULL)
 		return (NULL);
